backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
sqlalchemy==2.0.4
psycopg2-binary==2.9.5
PyMuPDF==1.21.1
pdfplumber==0.7.6
pytesseract==0.3.10
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
google-generativeai==0.3.1 